requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
pypdf
chromadb
python-dotenv